fix(#358): serialize stats-counter tests to eliminate xUnit parallel race
BlasManagedStatsTracker is a process-wide static. Tests that ClearCaches()
then assert specific counter values race with other concurrent tests.
xUnit's [Collection] attribute serializes tests in the same collection,
eliminating the race.

Tests moved from ScalarKernelTests.cs into a new StatsCounterTests.cs:
- BlasManaged_ClearCaches_ResetsCounters
- AutotuneDispatcher_CallSameShape10Times_LaterCallsHitCache
- BlasManaged_GetStats_TracksAutotuneHitsAndMisses
- BlasManaged_GetStats_DisableAutotune_DoesNotIncrementCounters
- JittedKernelCache_HitIncrementsStats
- BlasManaged_ClearCaches_ClearsJitCache
